rename filter: `initial -> initialize` to match finalize filter
