+++ cmd/ai-proxy/conf/routes.yml
@@ -5,7 +5,7 @@ routes:
     method: POST
     router: null
     filters:
-      - name: initial
+      - name: initialize
       - name: log-http
       - name: rate-limit
       - name: body-size-limit
